Add helper functions to and refactor util.
While writing tests, it became increasingly annoying to specify
the gridId for every call to the utility when testing the same
grid.  To make it easier, can now set a gridId on the utility so
that when testing the grid, the methods will grab the previously
specified gridId.  Refactored the method signatures to have gridId
as the last parameter so that it becomes optional.